+++ router.go
@@ -7,43 +7,17 @@ import (
 	"net/http"
 )
 
-func Router() *RootRouter {
+func Router(docsOptions ...*docs.Options) *RootRouter {
 	r := gin.Default()
+
+	docsOptions = append(docsOptions, &docs.Options{})
 
 	return &RootRouter{
 		routerGroup: routerGroup{
 			pathPrefix:   "",
 			rawRouter:    r,
 			middlewares:  nil,
-			Docs:         nil,
-			errorHandler: DefaultErrorHandler,
-		},
-		engine: r,
-	}
-}
-
-func DocumentedRouter(documentation *docs.Docs) *RootRouter {
-	r := gin.Default()
-
-	documentation.NewOpenAPI()
-
-	err := documentation.Valid()
-	if err != nil {
-		panic(err)
-	}
-
-	docHandler := docs.NewHandler(documentation)
-	docGroup := r.Group(documentation.OpenAPIPath)
-
-	docGroup.GET("", docHandler.Docs)
-	docGroup.GET("/openapi.json", docHandler.File)
-
-	return &RootRouter{
-		routerGroup: routerGroup{
-			pathPrefix:   "",
-			rawRouter:    r,
-			middlewares:  nil,
-			Docs:         documentation,
+			Docs:         docs.New(docsOptions[0]),
 			errorHandler: DefaultErrorHandler,
 		},
 		engine: r,
@@ -59,16 +33,12 @@ func (r *RootRouter) Engine() http.Handler {
 	return r.engine
 }
 
-func (r *RootRouter) Run(addr string) error {
+func (r *RootRouter) Run(host string, port string) error {
+	r.Docs.RegisterRoutes(r.rawRouter, port)
+
 	srv := &http.Server{
-		Addr:    addr,
+		Addr:    fmt.Sprintf("%s:%s", host, port),
 		Handler: r.engine,
-	}
-	if r.Docs != nil && !r.Docs.InMemory {
-		err := r.Docs.Build()
-		if err != nil {
-			panic(fmt.Sprintf("cannot build documentation; error: %v", err))
-		}
 	}
 
 	return srv.ListenAndServe()